requirements.txt
langchain-openai
langchain
python-dotenv
langchain_community
pandas
numpy
langchain_core
streamlit

